fix: remake ErrorScreen.js
